requirements.txt
# ── Core deep learning ──────────────────────────────────
ultralytics>=8.3.0
torch>=2.2.0
torchvision>=0.17.0

# ── ONNX export & quantization (Phase 5) ───────────────
onnx>=1.16.0
onnxruntime>=1.18.0

# ── Data handling ───────────────────────────────────────
roboflow>=1.1.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0
albumentations>=1.4.0

# ── Analysis & visualization ───────────────────────────
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
tqdm>=4.66.0

# ── Demo application (Phase 6) ─────────────────────────
streamlit>=1.35.0

# ── Utilities ───────────────────────────────────────────
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
